Some refactoring of path tree and element serialization for better maintainability.
